Redditors/agent/requirements.txt
praw==7.7.1
openai==1.3.0
supabase==2.0.3
python-dotenv==1.0.0
